[pm3] split classes to avoid circular dependencies
Having "Source" and "Target" in the same file but "Encoders" separately caused issues because Targets need Encoders which need Sources
